[#109] moved the colorPicker into common components
