+++ src/_common/colorPickerTooltip.js
