+++ src/_common/styles.css
@@ -1,0 +1,19 @@
+.tooltip {
+  position: absolute;
+  background: white;
+  visibility: hidden;
+}
+
+.tooltipButtonsWrp {
+  display: flex;
+  justify-content: space-between;
+}
+
+.tooltipButtonsWrp button~button {
+  margin-left: 0 !important;
+}
+
+.tooltipResult {
+  flex: 0.7;
+  background: red;
+}
