+++ src/column-limits/SettingsPage/htmlTemplates.js
@@ -42,24 +42,3 @@ export const columnTemplate = ({ columnId = '', dataGroupId = '', columnTitle = 
 
 export const dragOverHereTemplate = ({ dropzoneId = '', dropzoneClass = '' }) =>
   `<div  class="${style.addGroupDropzoneJH} ${dropzoneClass}" id="${dropzoneId}">Drag column over here to create group</div>`;
-
-export const colorPickerTooltipTemplate = ({
-  tooltipClass,
-  id,
-  colorPickerId,
-  colorPickerResultId,
-  okBtnId,
-  closeBtnId,
-  btnWrpClass,
-  colorPickerResultClass,
-}) => `
-  <div id="${id}" class="${tooltipClass}">
-    <div class="${btnWrpClass}">
-      <div class="${colorPickerResultClass}" id="${colorPickerResultId}"></div>
-      <div>
-        <button id="${okBtnId}" class="aui-button aui-button-primary" type="button">Ok</button>
-        <button id="${closeBtnId}" class="aui-button" type="button">Close</button>
-      </div>
-    </div>
-    <div id="${colorPickerId}"></div>
-  </div>`;
