+++ src/column-limits/SettingsPage/styles.css
@@ -82,24 +82,3 @@
 .formRightBlock {
   flex: 1;
 }
-
-
-.tooltip {
-  position: absolute;
-  background: white;
-  visibility: hidden;
-}
-
-.tooltipButtonsWrp {
-  display: flex;
-  justify-content: space-between;
-}
-
-.tooltipButtonsWrp button~button {
-  margin-left: 0 !important;
-}
-
-.tooltipResult {
-  flex: 0.7;
-  background: red;
-}
